requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
psycopg2-binary==2.9.6
requests==2.29.0
pandas==2.0.1
numpy==1.24.3
anthropic==0.3.6
python-multipart==0.0.6
fastapi-cors==0.0.6
jwt==1.3.1
markdown==3.4.3
aiofiles==23.1.0
